Extract DatabaseTestCase class with basic functionality for re-creating db for each test case
